+++ module-product-service/src/main/java/com/shop/productservice/product/controller/InternalProductController.java
@@ -1,9 +1,9 @@
-package com.shop.productservice.controller;
+package com.shop.productservice.product.controller;
 
-import com.shop.productservice.dto.response.ProductResponse;
-import com.shop.productservice.dto.response.ProductStockResponse;
-import com.shop.productservice.entity.Product;
-import com.shop.productservice.service.ProductService;
+import com.shop.productservice.product.dto.response.ProductResponse;
+import com.shop.productservice.product.dto.response.ProductStockResponse;
+import com.shop.productservice.product.entity.Product;
+import com.shop.productservice.product.service.ProductService;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
